requirements.txt
streamlit
faker
pandas
mysql-connector-python
plotly
python-dotenv